only keep one copy of test commands
